[Backend] SLight refactor data structure bidValue

+++ JDM-Legends/JDM-Legends-Dealership-Cars/src/main/java/com/jdm/legends/dealership/cars/service/repository/HistoryBidInterface.java
@@ -1,8 +1,10 @@
 package com.jdm.legends.dealership.cars.service.repository;
 
+import java.math.BigDecimal;
+
 public interface HistoryBidInterface {
     Integer getId();
 
-    Integer getBidValue();
+    BigDecimal getBidValue();
 
 }
